sync-mailgun: split api struct definitions in their own file
