+++ sync-mailgun/src/main.rs
@@ -1,63 +1,16 @@
+mod api;
+
 use std::cell::RefCell;
 use std::collections::{HashMap, HashSet};
 use std::env;
 use std::str;
 
+use crate::api::Empty;
 use curl::easy::{Easy, Form};
 use failure::{bail, format_err, Error, ResultExt};
 use rust_team_data::v1 as team_data;
 
 const DESCRIPTION: &str = "managed by an automatic script on github";
-
-mod api {
-    #[derive(serde_derive::Deserialize)]
-    pub struct ListResponse {
-        pub items: Vec<List>,
-        pub paging: Paging,
-    }
-
-    #[derive(serde_derive::Deserialize)]
-    pub struct RoutesResponse {
-        pub items: Vec<Route>,
-        pub total_count: usize,
-    }
-    #[derive(serde_derive::Deserialize)]
-    pub struct Route {
-        pub actions: Vec<String>,
-        pub expression: String,
-        pub id: String,
-        pub description: serde_json::Value,
-    }
-
-    #[derive(serde_derive::Deserialize)]
-    pub struct List {
-        pub access_level: String,
-        pub address: String,
-        pub members_count: u64,
-    }
-
-    #[derive(serde_derive::Deserialize)]
-    pub struct Paging {
-        pub first: String,
-        pub last: String,
-        pub next: String,
-        pub previous: String,
-    }
-
-    #[derive(serde_derive::Deserialize)]
-    pub struct MembersResponse {
-        pub items: Vec<Member>,
-        pub paging: Paging,
-    }
-
-    #[derive(serde_derive::Deserialize)]
-    pub struct Member {
-        pub address: String,
-    }
-}
-
-#[derive(serde_derive::Deserialize)]
-struct Empty {}
 
 fn main() {
     env_logger::init();
